requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-environ==0.10.0
django-taggit==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.4.0
mysqlclient==2.1.1
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
